SNOW-637237 Retry for 2 minutes since the first failure (#281)
Retry HTTP requests for some fixed total time instead of fixed number
of attempts. This provides better balance between retrying "long enough"
and not running out of memory.

1. If the server fails quickly, it is desirable to retry more times.
2. If the server fails slowly (times out after 1 minute, for example),
   many retry attempts can be cause the SDK to run out of memory
   due to the large amount of buffered data.

Retrying for N minutes makes is easier to reason about the amount
of data that is going to be buffered while waiting for the server
to recover.