SNOW-715618: Improve build latency for high throughput case with many channels (#321)
We're seeing some issue where the build time is high during high throughput case (20 million+ rows) with 10+ parallel channels going into the same table, and the root cause is that the buffer takes longer to transfer ownership and the next buffer already builds up with enough rows, so the file size becomes very big which is causing the compression time to be very high. In longer term, instead of checking the channel size, I think checking the allocator size per table would help this case. 

In this PR, it contains a few changes to improve the build time:

Fix a wrong buffer size logic which causes the flush to be scheduled unexpectedly
Use parallel stream to get the buffer data
Reduce the channel size limit to schedule the flush faster