NO-SNOW Fix flaky Windows test (#484)
The test `SnowflakeStreamingIngestChannelTest#testInsertRowThrottling` is
flaky on Windows. The reason is that we are using mockito to mock methods
from java.lang.Runtime in one thread and read them in another. These
changes are not always guaranteed to be visible in the other thread.

This PR introduces a new interface with two implementations to provide
memory info, so we don't need to mock anything anymore.