NO-SNOW: Limit the chunk size to avoid potential OOM issue in the Parquet Reader (#516)
Given that we have some limitations in the current Parquet Reader, this PR contains the change to limit the chunk size to avoid any potential OOM risks. It's currently set to 128MB uncompressed based on the reader expert and we also add the support to override it if needed to avoid any potential regressions.