SNOW-936900 Use upgraded version of snappy-java (#605)
The SDK is exporting an old version of its transitive dependency
snappy-java . This is visible in the new e2e jar project, which is
pulling in the snappy-java defined in hadoop-common and
not the version we define in dependencyManagement. This PR makes
snappy-java a direct dependency, so we can control the version.