NO-SNOW: Revert one change that updates public API for internal use case (#662)
Revert 2b702db

This change updates the public API for an internal special use case only which is not ideal, looks like we keep updating public APIs in #661 as well. Given that we're still in a POC phase, I suggest we do everything in another branch instead, hence removing this from the next public release and could discuss what's the best way to support this internal use case once the POC is done.