SNOW-2174149: Keep retrying on QUEUE_FULL instead of invalidating channels (#1022)
We have seen the issue in the O11y case when the chunk queue is full, channel write starts return QUEUE_FULL error, and it causes the channel to be invalidated. Channel invalidation results in a new set of channel reopen, which makes the issue even worse.

With this PR:
- We will keep retrying on QUEUE_FULL up to roughly 5mins instead of just a few seconds before invalidating the channel.
- Add a limit on the registration queue and throttle insertRows once it reaches a certain size
- Both parameters are configurable in case there is any regressions.